[DOP-3911]: More refactoring, update tests to use set
