+++ src/monorepo/index.ts
@@ -1,5 +1,6 @@
-import { getProjectDirFromPath, getProjectDirFromPathSet, getSnootyDirSet } from './utils/paths';
+import { getSnootyDirSet } from './utils';
 import { GitCommitInfo } from './types/github-types';
+import { getProjectDirFromPathSet } from './services/get-paths';
 
 interface FileUpdatePayload {
   repoName: string;
@@ -19,8 +20,8 @@ export async function getMonorepoPaths(fileUpdates: FileUpdatePayload): Promise<
 
   const snootyDirSet = await getSnootyDirSet(commitInfo);
 
-  const projects = await Promise.all(updatedFilePaths.map((path) => getProjectDirFromPath(path, commitInfo)));
-  // const projects = await Promise.all(updatedFilePaths.map((path) => getProjectDirFromPathSet(path, snootyDirSet)));
+  // const projects = await Promise.all(updatedFilePaths.map((path) => getProjectDirFromPath(path, commitInfo)));
+  const projects = updatedFilePaths.map((path) => getProjectDirFromPathSet(path, snootyDirSet));
 
   // remove empty strings and remove duplicated values
   return Array.from(new Set(projects.filter((dir) => !!dir)));
